refactor: delete bootstrap and rework country component
