+++ src/Routes.tsx
@@ -1,12 +1,12 @@
 import { Route, Switch } from 'react-router-dom';
-import Country from './pages/Country';
+import CountryPage from './pages/CountryPage';
 import MainPage from './pages/MainPage';
 
 const Routes = (): JSX.Element => (
   <Switch>
     <Route path="/" exact component={MainPage} />
     {/* <Route path="/country/:id" exact component={Country} /> */}
-    <Route path="/country/" exact component={Country} />
+    <Route path="/country/" exact component={CountryPage} />
   </Switch>
 );
