+++ src/index.tsx
@@ -3,8 +3,6 @@ import ReactDOM from 'react-dom';
 import { BrowserRouter } from 'react-router-dom';
 import App from './App';
 
-import './index.css';
-
 const renderApp = () => (
   <BrowserRouter>
     <App />
