+++ src/pages/CountryPage.tsx
@@ -1,0 +1,8 @@
+// import { useParams } from 'react-router-dom';
+import Country from '../components/Country/Country';
+
+const CountryPage: React.FC = () => {
+  // const { id } = useParams();
+  return <Country />;
+};
+export default CountryPage;
